apply code formatting
The project was source-formatted using Lazarus IDE
(which in turn invokes JEDI code formatting).

Settings file added to repo (default settings actually).

Considering the benefits and the drawbacks, I thought it's worth doing.
Downside is the merge process in some cases,
benefits include maintainability and readability (in case of >1 developers).